requirements.txt
disnake==2.9.0
requests==2.31.0
aiosqlite==0.19.0
openai==1.12.0
python-dotenv==1.0.0